changed stats to be floats instead of ints, aaaand a lot of other things to be floats

+++ Assets/_Scripts/Battle/BattlePickup.cs
@@ -111,7 +111,7 @@ namespace Lis
             _hero.Pull.OnValueChanged -= SetPickUpRadius;
         }
 
-        void SetPickUpRadius(int i)
+        void SetPickUpRadius(float i)
         {
             _sphereCollider.radius = i;
         }
